prevent duplication of set matrix from Array code
