+++ src/gui/scene/AnchoredGroup.js
@@ -1,7 +1,8 @@
 import * as THREE from "../../../thirdPartyCode/three/three.module.js"
+import {setMatrixFromArray} from "./utils.js"
 
 /**
- * @typedef {number[]} MatrixAsArray - a 4x4 matrix representated as an column-mayor array of 16 numbers 
+ * @typedef {import("./utils.js").MatrixAsArray} MatrixAsArray
  */
 
 /**
@@ -28,7 +29,7 @@ class AnchoredGroup {
      * @param {MatrixAsArray} array 
      */
     setMatrixFromArray(array) {
-        AnchoredGroup.setMatrixFromArray(this.#group.matrix, array)
+        setMatrixFromArray(this.#group.matrix, array)
     }
 
     /**
@@ -61,19 +62,6 @@ class AnchoredGroup {
      */
     getInternalObject() {
         return this.#group;
-    }
-
-     /**
-      * small helper function for setting three.js matrices from the custom format we use
-      * @param {THREE.Matrix} matrix
-      * @param {MatrixAsArray} array
-      */
-     static setMatrixFromArray(matrix, array) {
-        matrix.set( array[0], array[4], array[8], array[12],
-            array[1], array[5], array[9], array[13],
-            array[2], array[6], array[10], array[14],
-            array[3], array[7], array[11], array[15]
-        );
     }
 }
